+++ crates/ruma-events/src/policy/rule/user.rs
@@ -6,7 +6,7 @@ use ruma_macros::EventContent;
 use serde::{Deserialize, Serialize};
 
 use super::{PolicyRuleEventContent, PossiblyRedactedPolicyRuleEventContent};
-use crate::{EventContent, PossiblyRedactedStateEventContent, StateEventType};
+use crate::{EventContent, PossiblyRedactedStateEventContent, StateEventType, StaticEventContent};
 
 /// The content of an `m.policy.rule.user` event.
 ///
@@ -33,4 +33,8 @@ impl EventContent for PossiblyRedactedPolicyRuleUserEventContent {
 
 impl PossiblyRedactedStateEventContent for PossiblyRedactedPolicyRuleUserEventContent {
     type StateKey = String;
+}
+
+impl StaticEventContent for PossiblyRedactedPolicyRuleUserEventContent {
+    const TYPE: &'static str = PolicyRuleUserEventContent::TYPE;
 }